server/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
pillow==10.3.0